requirements.txt
aio-api-ros==0.0.19
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
librouteros==3.4.1
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
netmiko==4.5.0
ntc_templates==7.8.0
paramiko==3.5.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
RouterOS-api==0.21.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
textfsm==1.1.3
toml==0.10.2
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.19.0
